Restructured opengl.mesh.selection
