+++ amulet_map_editor/api/opengl/mesh/selection/__init__.py
@@ -1,5 +1,2 @@
-from .render_selection import RenderSelection
-from .render_selection_highlightable import RenderSelectionHighlightable
-from .render_selection_editable import RenderSelectionEditable
-from .render_selection_group import RenderSelectionGroup
-from .render_selection_group_editable import RenderSelectionGroupEditable
+from .box import RenderSelection, RenderSelectionEditable, RenderSelectionHighlightable
+from .group import RenderSelectionGroup, RenderSelectionGroupEditable
